widden-question/requirements.txt
# Core dependencies for widden-question project

# Wikipedia API
wikipediaapi>=0.6.0

# Graph analysis
networkx>=3.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0

# Database
pymongo>=4.0.0

# Configuration
python-dotenv>=1.0.0

# Optional (for pandas compatibility if needed)
pandas>=2.0.0

# Progress bar (optional, for scraper)
tqdm>=4.65.0
